third_party/WebKit/LayoutTests/http/tests/websocket/bufferedAmount-after-close-in-busy-expected.txt

CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
CONSOLE ERROR: WebSocket is already in CLOSING or CLOSED state.
WebSocket bufferedAmount after closed in busy

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS bufferedAmountBeforeClose + closeFrameSize >= bufferedAmountAfterClose is true
Closed.
PASS ws.readyState is 3
PASS ws.bufferedAmount <= bufferedAmountAfterClose is true
Testing send(string)...
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 21
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 0
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 1
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 125
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 126
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 65535
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 65536
Testing send(ArrayBuffer)...
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 0
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 1
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 125
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 126
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 65535
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 65536
Testing send(Blob)...
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 0
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 1
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 125
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 126
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 65535
PASS ws.send(messageToSend) is undefined.
PASS bufferedAmountDifference is 65536
PASS successfullyParsed is true

TEST COMPLETE
